ENH: add support for specifying dtypes (#24)
* ENH: better errors relating to float issues

* ENH: add dtype utils

* ENH: allow converting samples with dtypes

* TST: dtype utils tests

* ENH: specify dtype throughout code

* TST: add dtypes to tests

+++ src/aspire/samplers/importance.py
@@ -8,7 +8,11 @@ class ImportanceSampler(Sampler):
     def sample(self, n_samples: int) -> Samples:
         x, log_q = self.prior_flow.sample_and_log_prob(n_samples)
         samples = Samples(
-            x, log_q=log_q, xp=self.xp, parameters=self.parameters
+            x,
+            log_q=log_q,
+            xp=self.xp,
+            parameters=self.parameters,
+            dtype=self.dtype,
         )
         samples.log_prior = samples.array_to_namespace(self.log_prior(samples))
         samples.log_likelihood = samples.array_to_namespace(
